src/requirements.txt
boto3==1.22.2
requests==2.27.1
python-dotenv==0.20.0
